refactor: improve document metadata handling and usage of default values
